Move ExperimentController to use permission classes instead of ACL bits
Fix parent/child sample filtering, which should fix the SampleSet test

9250 : **Parent column processing leaves incomplete Derive experiment, causes NPE

+++ experiment/src/org/labkey/experiment/api/AbstractProtocolOutputImpl.java
@@ -108,7 +108,7 @@ public abstract class AbstractProtocolOutputImpl<Type extends ProtocolOutput> ex
         {
             return null;
         }
-        _sourceApp = ExperimentServiceImpl.get().getExpProtocolApplication(_object.getSourceApplicationId());
+        _sourceApp = ExperimentServiceImpl.get().getExpProtocolApplication(_object.getSourceApplicationId().intValue());
         return _sourceApp;
     }
 
@@ -141,7 +141,7 @@ public abstract class AbstractProtocolOutputImpl<Type extends ProtocolOutput> ex
         {
             return null;
         }
-        return ExperimentServiceImpl.get().getExpRun(_object.getRunId());
+        return ExperimentServiceImpl.get().getExpRun(_object.getRunId().intValue());
     }
 
     public void setSourceApplication(ExpProtocolApplication app)
